separation of responsibility
